Classify duplicate and overlapping zones as configuration errors
